# api/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/audio/focus/{path}": {
      "get": {
        "operationId": "BuiltinAudioController_getFocusAudio",
        "parameters": [
          {
            "name": "path",
            "required": true,
            "in": "path",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "BuiltinAudio"
        ]
      }
    },
...
    "/auth/me": {
      "get": {
        "operationId": "getAuthSession",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Auth"
        ]
      },
      "patch": {
        "operationId": "updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Auth"
        ]
      },
      "delete": {
        "operationId": "AuthController_deleteAccount",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password": {
      "post": {
        "operationId": "changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/auth/google/register": {
      "post": {
        "operationId": "AuthController_googleRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleRegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/oauth/exchange": {
      "post": {
        "operationId": "AuthController_oauthExchange",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OAuthExchangeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "iTu API",
    "description": "Authoritative REST contract for iTu productivity & learning system",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {}
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {}
      },
      "DeleteAccountDto": {
        "type": "object",
        "properties": {}
      },
      "GoogleRegisterDto": {
        "type": "object",
        "properties": {}
      },
      "OAuthExchangeDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}